requirements.txt
python-telegram-bot
instagrapi
yt-dlp
ShazamAPI
librosa
requests
python-dotenv
